feat: opendde defaults to a bfloat16 trunk with a float32 diffusion module
The port shipped upstream's float32 trunk. On this hardware that default
dies near 1,400 residues, so the shipped configuration cannot run the sizes
the port is for, while the bf16 arm the port already had sat behind a flag.

It is not a precision trade anyone has to weigh. On all eight panel cases the
bf16 arm is 13-39% faster and 33-52% lighter (7ST3: 80.27 -> 52.57 s and
14204 -> 6890 MiB), its coordinates sit 0.014-0.126 A from the fp32 arm
against within-set spreads of 0.33-10.7 A, and measured against the same
upstream rows it is no further away than fp32 — marginally closer on three of
the four cases read.

The cast is already the shape AlphaFold 3 ships and was not chosen to be:
`cast_trunk_params` narrows `input_embedder`, `pairformer_output`,
`structural_expander` and `structural_refiner`, and leaves the diffusion
module, the distogram and the confidence head float32. `structural_refiner`
is the second Pairformer trunk, not the denoiser.

`--trunk-dtype fp32` pins upstream's policy back, and the arena preflight now
describes fp32 as the pinned choice rather than the default. One stale
comment fixed with it: the preflight read `None` as "left at the released
float32", which the flip inverts.

CPU: 603 passed across tests/models/opendde, tests/test_cache.py and
tests/test_backends.py; ruff clean.

+++ src/foldjax/backends/opendde.py
@@ -82,7 +82,7 @@ _RELEASED_COMPILE_DEFAULTS: dict[str, object] = {
     "diffusion_attention_backend": "xla_jit",
     "trunk_single_attention_backend": "xla_jit",
     "structural_single_attention_backend": "xla_jit",
-    "trunk_dtype": "fp32",
+    "trunk_dtype": "bf16",
     "chunk_policy": "auto",
     "cp_devices": 1,
     "cp_layout": "auto",
